[IMP] Improve retrofit

+++ app/src/main/java/com/example/homescreen/nutrition/FoodAPI.kt
@@ -2,15 +2,16 @@ package com.example.homescreen.nutrition
 
 data class FoodAPI(
     val name: String,
-    val calories: Int,
-    val serving_size_g: Int,
-    val fat_total_g: Int,
-    val fat_saturated_g: Int,
-    val protein_g: Int,
+    val calories: Double,
+    val serving_size_g: Double,
+    val fat_total_g: Double,
+    val fat_saturated_g: Double,
+    val protein_g: Double,
     val sodium_mg: Int,
     val potassium_mg: Int,
     val cholesterol_mg: Int,
-    val carbohydrates_total_g: Int,
-    val fiber_g: Int,
-    val sugar_g: Int
+    val carbohydrates_total_g: Double,
+    val fiber_g: Double,
+    val sugar_g: Double
 )
+
